Thêm chức năng quản lý sản phẩm: bổ sung ProductController, IProductService và ProductService; cập nhật các thực thể liên quan và cấu trúc cơ sở dữ liệu cho kích cỡ và màu sắc sản phẩm.

+++ Core/Entities/Shoppingcart.cs
@@ -20,7 +20,7 @@ public class Shoppingcart
     /// <summary>
     ///     Tổng giá trị hiện tại của giỏ hàng
     /// </summary>
-    public decimal TotalPrice { get; set; }
+    public int TotalPrice { get; set; }
 
     /// <summary>
     ///     Người cập nhật cuối cùng
